--- a/python/geoipsets.conf
+++ b/python/geoipsets.conf
@@ -9,12 +9,12 @@ provider=dbip,maxmind
 # valid values are: 'iptables', 'nftables'
 # iptables: builds 'ipset' compatible sets
 # nftables: builds nftables compatible sets
-# if the property doesn't exist, or exists but is empty, nftables sets are generated (default)
+# default: nftables
 firewall=iptables,nftables
 
 # list of IP protocols to build sets for
 # valid values are: 'ipv4', 'ipv6'
-# if the property doesn't exist or exists, but is empty, ipv4 sets will be generated (default)
+# default: ipv4
 address-family=ipv4,ipv6
 
 # specify which countries to build sets for