When only the `nftables` firewall type is selected, `geoipsets` will avoid converting `dbip` data to CIDR notation and simply pass the IP ranges directly to `nftables` without modification. This results in a ~50% performance improvement.
```
# before
[tags/v2.3.2] % time python -m geoipsets --provider dbip --firewall nftables --config-file ~/geoipsets.conf
Building geoipsets...
python -m geoipsets --provider dbip --firewall nftables --config-file   11.66s user 0.07s system 94% cpu 12.387 total

# after
[dbip-perf] % time python -m geoipsets --provider dbip --firewall nftables --config-file ~/geoipsets.conf
Building geoipsets...
python -m geoipsets --provider dbip --firewall nftables --config-file   5.07s user 0.07s system 88% cpu 5.789 total
```
Fixes #25 

Also includes a small change to only delete existing sets if the current run will re-create them. Otherwise they are left untouched. This brings the behaviour of the `dbip` provider inline with the `MaxMind` provider.